Revert "Remove unused bool operation, add note about extra accumulator var"
This reverts commit 2ad7b7b915efc3c895157d69ce36a25eedf6a57f.

I mistakenly intuited that it wasn't necessary to compute the disjunction of the
flags and pass them on. @coancea pointed out that the `scan` operation is
parallel, and that the disjuncted flag will be passed on and used in subsequent
calls to the function. In other words, the `scan` isn't a `scanl`...